backend/pyproject.toml
[tool.poetry]
name = "arcade_rocket_approval"
version = "0.0.2"
description = "Rocket Mortgage agent for guiding user through Rocket mortgage load application"
authors = ["BlakeItOut <blakeshaw@rocketmortgage.com>", "Spartee <sam@arcade.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain-arcade = "^1.3.0"
arcade-ai = "^1.1.2"
arcadepy = "^1.3.1"
langchain = ">=0.3.0"
langchain_core = ">=0.3.0"
langchain_openai = ">=0.2.0"
python-dotenv = "^1.0.1"
google-api-core = "2.19.1"
google-api-python-client = "2.137.0"
google-auth = "2.32.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.1"
googleapis-common-protos = "1.63.2"
trustcall = "0.0.38"


[tool.poetry.dev-dependencies]
pytest = "^8.3.0"
pytest-cov = "^4.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ruff = "^0.7.4"
langgraph-cli = { version = ">=0.1.71", extras = ["inmem"] }


[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["arcade_rocket_approval/**/*.py"]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true
